Commit for today 10/29/20 - Aman
Few file uploads. Please download the previous commit if you want *working* code this is broken for some reason, i'll update as i fix it. the run_test.sh has some comments that you should reference about next steps if you want to contribute. The TestKafka/TestConsumer are just test kafka files so you can see on a basic level how things run. You run the TestConsumer first and then the TestKafka for it to generate some output. mock_process/collect need to be updated because they're both running when I just run mock_collect which is not right lmao. Not commiting this to main so it's easy to find for later.